Move intents into separate files
