+++ ask_fandom/intents/semantic_media_wiki/__init__.py
@@ -1,0 +1,7 @@
+"""
+SemanticMediaWiki based intents
+"""
+from .base import SemanticFandomIntent
+
+from .tv_series import EpisodeFactIntent, PersonFactIntent
+from. wowwiki import WoWGroupsMemberIntent
