+++ ask_fandom/intents/semantic_media_wiki/tv_series.py
@@ -1,8 +1,7 @@
 """
-Specific intents
+Intent handling episodes and characters questions
 """
-# pylint: disable=too-few-public-methods
-from ask_fandom.intents.smw import SemanticFandomIntent
+from .base import SemanticFandomIntent
 
 
 class PersonFactIntent(SemanticFandomIntent):
@@ -52,27 +51,5 @@ class EpisodeFactIntent(SemanticFandomIntent):
 
         site = self.get_mw_client('tardis.fandom.com')
         values = self.get_smw_property_for_page(site, name, prop)
-
-        return values[0] if values else None
-
-
-class WoWGroupsMemberIntent(SemanticFandomIntent):
-    """
-    Provides World Of Warcraft nations membership data
-    """
-    ANSWER_PHRASE = '{name} is a member of "{answer}" {group}.'
-
-    @property
-    def _answer(self):
-        # https://wowwiki.fandom.com/wiki/Special:Browse/Alterac
-        # ({'name': 'the Alterac', 'group': 'faction'}
-        name = str(self.args['name'])
-
-        # remove 'the'
-        if name.lower().startswith('the '):
-            self.args['name'] = name = name[4:]
-
-        site = self.get_mw_client('wowwiki.fandom.com')
-        values = self.get_smw_property_for_page(site, name, self.args['group'])
 
         return values[0] if values else None
